Update vc_image_convert tool. (#859)
* Update vc_image_convert tool.

+++ tools/utilities/vc_image_convert/sof_convert.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 OUTPUT_PATH=./output_files
-SOF_FILE=./input_files/chip.sof
+FACTORY_SOF_FILE=./input_files/vc_factory.sof
+USER_SOF_FILE=./input_files/vc_user.sof
 COF_FILE=vcp_gen_pof.cof
 DTB_FILE=./input_files/vc_dtb.bin
 
@@ -13,41 +14,50 @@ do
 		break
 	fi
 done < $COF_FILE
-ext=`date +%Y-%m-%d_%H:%M:%S`.bin
-dest_file=${pof_file##*/}
-bin_file=${dest_file%.*}'.bin'
+bin_file=vc.bin
+dest_file=vc_flash.bin
 dest_reversed_file=vc_reverse.bin
 dest_reversed_lite_file=vc_reverse_flash.bin
-dest_file=${dest_file%.*}'_'$ext
 
-while getopts f:u:d:o:h opt
+while getopts f:u:d:o:vh opt
 do
 	case "$opt" in
 	f) factory_sof=$OPTARG;;
 	u) user_sof=$OPTARG;;
 	d) dtb_bin=$OPTARG;;
 	o) dest_file=$OPTARG;;
-	h) echo -e "Usage: $0 -u [sof_file] -o [bin_file]\n\t-f\tSet factory image file\n\t-u\tSet user image file\n\t-d\tSet device tree table file name\n\t-o\tSet output file name";exit 0;;
+	h) echo -e "Usage: $0 -u [sof_file] -o [bin_file]\n\t-f\tSet factory image file\n\t-u\tSet user image file\n\t-d\tSet device tree table file name\n\t-o\tSet output file name\n\t-v\tShow publish date of this tool";exit 0;;
+	v) echo "2019.05.23";exit 0;;
 	*) exit 0;;
 	esac
 done
 
 if [ ! -n "$user_sof" ]
 then
-	user_sof=$SOF_FILE
+	user_sof=$USER_SOF_FILE
+fi
+if [ ! -f "$user_sof" ]
+then
+	echo $user_sof "not found, please try -u option to select one"
+	exit 0
 fi
 if [ ! -n "$factory_sof" ]
 then
-	factory_sof=$user_sof
+	factory_sof=$FACTORY_SOF_FILE
 fi
 if [ ! -f "$factory_sof" ]
 then
-	echo $factory_sof "not found"
+	echo $factory_sof "not found, please try -f option to select one"
 	exit 0
 fi
 if [ ! -n "$dtb_bin" ]
 then
 	dtb_bin=$DTB_FILE
+fi
+if [ ! -f "$dtb_bin" ]
+then
+	echo $dtb_bin "not found, please try -d option to select one"
+	exit 0
 fi
 
 sed "s!factory_image.sof!$factory_sof!" $COF_FILE > $cof_file
